api/django/requirements.txt
django==3.0.2
pymysql
pandas
pillow
tensorflow
mtcnn
numpy
requests
opencv-python-headless